Move expiry field to the dedicated file
